split sections up a bit
