+++ example/src/pages/leaderboard/StatisticsSection.tsx
@@ -1,0 +1,45 @@
+import { useQuery } from "convex/react";
+import { api } from "../../../convex/_generated/api";
+import { Text, Card, Stack, Group, Title } from "@mantine/core";
+
+export function StatisticsSection() {
+  const totalCount = useQuery(api.leaderboard.countScores);
+  const totalSum = useQuery(api.leaderboard.sumNumbers);
+
+  return (
+    <Group gap="md">
+      <Card bg="dark.7" p="lg" style={{ flex: 1 }}>
+        <Stack gap="xs">
+          <Text size="sm" c="gray.4">
+            Total Scores
+          </Text>
+          <Title order={2} c="white">
+            {totalCount ?? "Loading..."}
+          </Title>
+        </Stack>
+      </Card>
+      <Card bg="dark.7" p="lg" style={{ flex: 1 }}>
+        <Stack gap="xs">
+          <Text size="sm" c="gray.4">
+            Total Sum
+          </Text>
+          <Title order={2} c="white">
+            {totalSum ?? "Loading..."}
+          </Title>
+        </Stack>
+      </Card>
+      <Card bg="dark.7" p="lg" style={{ flex: 1 }}>
+        <Stack gap="xs">
+          <Text size="sm" c="gray.4">
+            Average Score
+          </Text>
+          <Title order={2} c="white">
+            {totalCount && totalSum
+              ? (totalSum / totalCount).toFixed(2)
+              : "Loading..."}
+          </Title>
+        </Stack>
+      </Card>
+    </Group>
+  );
+}
